팀 프로젝트 2주차 코드 리뷰 요청

See merge request sw_track/class_05/web_project/team03/back-end!81

+++ src/services/adminService.js
@@ -1,19 +1,23 @@
 const { productModel } = require('../db/models');
 const { orderModel } = require('../db/models');
 const { categoryModel } = require('../db/models');
+const { userModel } = require('../db/models')
+const AppError = require('../misc/AppError');
 
 class AdminService {
-  constructor(productModel, orderModel, categoryModel) {
+  constructor(productModel, orderModel, categoryModel, userModel) {
     this.productModel = productModel;
     this.orderModel = orderModel;
     this.categoryModel = categoryModel;
+    this.userModel = userModel;
   }
 
   // 상품 등록 관리자
-  async addProduct(product) {
-    const { id, name, categoryId, price, subImage, keyWord, description, mainImage } = product;
-    if (!id || !name || !categoryId || !price || !subImage || !keyWord || !description || !mainImage) {
-      throw new Error('필수 정보를 모두 입력해주세요.');
+  async postProduct(product) {
+    const { id } = product;
+    const Product = await this.productModel.findOne(id);
+    if (Product) {
+      throw new AppError('Bad Request', 400, '이미 존재하는 상품입니다.');
     }
     const result = await this.productModel.create(product);
     return result;
@@ -22,14 +26,18 @@ class AdminService {
   // 상품 단건 조회 관리자
   async getProduct(productId) {
     const product = await this.productModel.findOne(productId);
+    if (!product) {
+      throw new AppError('Bad Request', 400, '존재하지 않는 상품입니다.');
+    }
     return product;
   }
 
   // 상품 단건 수정 관리자
   async putProduct(product) {
-    const { id, name, categoryId, price, subImage, keyWord, description, mainImage } = product;
-    if (!id || !name || !categoryId || !price || !subImage || !keyWord || !description || !mainImage) {
-      throw new Error('필수 정보를 모두 입력해주세요.');
+    const { id } = product;
+    const Product = await this.productModel.findOne(id);
+    if (!Product) {
+      throw new AppError('Bad Request', 400, '존재하지 않는 상품입니다.');
     }
     const result = await this.productModel.update(product);
     return result;
@@ -37,6 +45,11 @@ class AdminService {
 
   // 상품 단건 삭제 관리자
   async deleteProduct(productId) {
+    const { id } = product;
+    const Product = await this.productModel.findOne(id);
+    if (!Product) {
+      throw new AppError('Bad Request', 400, '존재하지 않는 상품입니다.');
+    }
     const result = await this.productModel.deleteOne(productId);
     return result;
   }
@@ -49,25 +62,33 @@ class AdminService {
 
   // 주문 내역 특정 사용자 조회 관리자
   async getOrder(id) {
-    const order = await this.orderModel.findOrder(id);
+    const user = await this.userModel.findbyId(id);
+    if (!user) {
+      throw new AppError('Bad Request', 400, '존재하지 않는 사용자입니다.');
+    }
+    const order = await this.orderModel.findOrdersByUserId(id);
     return order;
   }
 
   // 주문 내역 삭제 관리자
   async deleteOrder(id) {
+    const user = await this.userModel.findbyId(id);
+    if (!user) {
+      throw new AppError('Bad Request', 400, '존재하지 않는 사용자입니다.');
+    }
     const result = await this.orderModel.deleteOrder(id);
     return result;
   }
   //카테고리 추가
-  async addCategory(categoryInfo) {
+  async postCategory(categoryInfo) {
     const { id, name } = categoryInfo;
     const category = await this.categoryModel.findById(id);
     if (category) {
-      throw new Error('이미 존재하는 카테고리ID입니다.');
+      throw new AppError('Bad Request', 400, '이미 존재하는 카테고리ID입니다.');
     }
     const category2 = await this.categoryModel.findByName(name);
     if (category2) {
-      throw new Error('이미 존재하는 카테고리 이름입니다.');
+      throw new AppError('Bad Request', 400, '이미 존재하는 카테고리 이름입니다.');
     }
     const result = await this.categoryModel.create({ id, name });
     return result;
@@ -76,7 +97,7 @@ class AdminService {
   async getCategory(id) {
     const category = await this.categoryModel.findById(id);
     if (!category) {
-      throw new Error('존재하지 않는 카테고리입니다.');
+      throw new AppError('Bad Request', 400, '존재하지 않는 카테고리입니다.');
     }
     const result = await this.categoryModel.findById(id);
     return result;
@@ -86,7 +107,7 @@ class AdminService {
     const { id, name } = categoryInfo;
     const category = await this.categoryModel.findById(id);
     if (!category) {
-      throw new Error('존재하지 않는 카테고리입니다.');
+      throw new AppError('Bad Request', 400, '존재하지 않는 카테고리입니다.');
     }
     const result = await this.categoryModel.update(categoryInfo);
     return result;
@@ -96,7 +117,7 @@ class AdminService {
   async deleteCategory(id) {
     const category = await this.categoryModel.findById(id);
     if (!category) {
-      throw new Error('존재하지 않는 카테고리입니다.');
+      throw new AppError('Bad Request', 400, '존재하지 않는 카테고리입니다.');
     }
     const result = await this.categoryModel.deleteCategory(id);
     return result;
@@ -104,6 +125,10 @@ class AdminService {
 
   // 주문 내역 주문상태 변경 관리자
   async putStatus(id, status) {
+    const order = await this.orderModel.findOrderByOrderId(id);
+    if (!order) {
+      throw new AppError('Bad Request', 400, '존재하지 않는 주문입니다.')
+    }
     const result = await this.orderModel.updateStatus(id, status);
     return result;
   }
